Move MeasurementExchange to ag_data/utilities.py
