+++ ag_data/utilities.py
@@ -1,6 +1,28 @@
+import attr
+
+from django.utils import timezone
+
 from ag_data import models
 from ag_data.presets.built_in_content import built_in_sensor_types
 
+@attr.s
+class MeasurementExchange(object):
+    event = attr.ib(kw_only=True, validator=attr.validators.instance_of(models.AGEvent))
+    timestamp = attr.ib(
+        kw_only=True, validator=attr.validators.instance_of(timezone.datetime)
+    )
+    sensor = attr.ib(
+        kw_only=True, validator=attr.validators.instance_of(models.AGSensor)
+    )
+    reading = attr.ib(kw_only=True, validator=attr.validators.instance_of(dict))
+
+    @property
+    def sensor_type(self):
+        return self.sensor.type_id
+
+    @property
+    def processing_formula(self):
+        return self.sensor_type.processing_formula
 
 def createOrResetAllBuiltInSensorTypes():
     for index in range(len(built_in_sensor_types)):
